% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExportBaselineSurveyResponses.R
\docType{class}
\name{ExportBaselineSurveyResponses}
\alias{ExportBaselineSurveyResponses}
\title{ExportBaselineSurveyResponses}
\format{
An R6 class object.
}
\description{
ExportBaselineSurveyResponses class
}
\details{
Prepare data export of baseline survey responses.
}
\section{Super classes}{
\code{\link[exceedapi:SnapshotBackend]{exceedapi::SnapshotBackend}} -> \code{\link[exceedapi:API]{exceedapi::API}} -> \code{\link[exceedapi:Step]{exceedapi::Step}} -> \code{\link[exceed.pipelines:LoadSurveyResponses]{exceed.pipelines::LoadSurveyResponses}} -> \code{\link[exceed.pipelines:LoadBaselineSurveyResponses]{exceed.pipelines::LoadBaselineSurveyResponses}} -> \code{ExportBaselineSurveyResponses}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-transform}{\code{ExportBaselineSurveyResponses$transform()}}
\item \href{#method-clone}{\code{ExportBaselineSurveyResponses$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="exceedapi" data-topic="Step" data-id="add_step">}\href{../../exceedapi/html/Step.html#method-add_step}{\code{exceedapi::Step$add_step()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="exceedapi" data-topic="Step" data-id="add_step_object">}\href{../../exceedapi/html/Step.html#method-add_step_object}{\code{exceedapi::Step$add_step_object()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="exceedapi" data-topic="Step" data-id="inverse_transform">}\href{../../exceedapi/html/Step.html#method-inverse_transform}{\code{exceedapi::Step$inverse_transform()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="exceedapi" data-topic="Step" data-id="progress_bar">}\href{../../exceedapi/html/Step.html#method-progress_bar}{\code{exceedapi::Step$progress_bar()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="exceedapi" data-topic="Step" data-id="run">}\href{../../exceedapi/html/Step.html#method-run}{\code{exceedapi::Step$run()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="exceed.pipelines" data-topic="LoadSurveyResponses" data-id="initialize">}\href{../../exceed.pipelines/html/LoadSurveyResponses.html#method-initialize}{\code{exceed.pipelines::LoadSurveyResponses$initialize()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="exceed.pipelines" data-topic="LoadBaselineSurveyResponses" data-id="get_responses">}\href{../../exceed.pipelines/html/LoadBaselineSurveyResponses.html#method-get_responses}{\code{exceed.pipelines::LoadBaselineSurveyResponses$get_responses()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-transform"></a>}}
\if{latex}{\out{\hypertarget{method-transform}{}}}
\subsection{Method \code{transform()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExportBaselineSurveyResponses$transform(.data, .collect, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExportBaselineSurveyResponses$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
