% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExportPhenotypes.R
\docType{class}
\name{ExportPhenotypes}
\alias{ExportPhenotypes}
\title{ExportPhenotypes}
\format{
An R6 class object.
}
\description{
ExportPhenotypes class
}
\details{
Export phenotypes defined by Leicester GenEpi group.
}
\section{Super classes}{
\code{\link[exceedapi:SnapshotBackend]{exceedapi::SnapshotBackend}} -> \code{\link[exceedapi:API]{exceedapi::API}} -> \code{\link[exceedapi:Step]{exceedapi::Step}} -> \code{\link[exceed.pipelines:ExportTable]{exceed.pipelines::ExportTable}} -> \code{ExportPhenotypes}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-transform}{\code{ExportPhenotypes$transform()}}
\item \href{#method-prepare_metadata}{\code{ExportPhenotypes$prepare_metadata()}}
\item \href{#method-prepare_dataset}{\code{ExportPhenotypes$prepare_dataset()}}
\item \href{#method-clone}{\code{ExportPhenotypes$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="exceedapi" data-topic="Step" data-id="add_step">}\href{../../exceedapi/html/Step.html#method-add_step}{\code{exceedapi::Step$add_step()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="exceedapi" data-topic="Step" data-id="add_step_object">}\href{../../exceedapi/html/Step.html#method-add_step_object}{\code{exceedapi::Step$add_step_object()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="exceedapi" data-topic="Step" data-id="initialize">}\href{../../exceedapi/html/Step.html#method-initialize}{\code{exceedapi::Step$initialize()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="exceedapi" data-topic="Step" data-id="inverse_transform">}\href{../../exceedapi/html/Step.html#method-inverse_transform}{\code{exceedapi::Step$inverse_transform()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="exceedapi" data-topic="Step" data-id="progress_bar">}\href{../../exceedapi/html/Step.html#method-progress_bar}{\code{exceedapi::Step$progress_bar()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="exceedapi" data-topic="Step" data-id="run">}\href{../../exceedapi/html/Step.html#method-run}{\code{exceedapi::Step$run()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="exceed.pipelines" data-topic="ExportTable" data-id="add_prefix">}\href{../../exceed.pipelines/html/ExportTable.html#method-add_prefix}{\code{exceed.pipelines::ExportTable$add_prefix()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="exceed.pipelines" data-topic="ExportTable" data-id="add_shared_metadata">}\href{../../exceed.pipelines/html/ExportTable.html#method-add_shared_metadata}{\code{exceed.pipelines::ExportTable$add_shared_metadata()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="exceed.pipelines" data-topic="ExportTable" data-id="alert_danger">}\href{../../exceed.pipelines/html/ExportTable.html#method-alert_danger}{\code{exceed.pipelines::ExportTable$alert_danger()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="exceed.pipelines" data-topic="ExportTable" data-id="alert_warning">}\href{../../exceed.pipelines/html/ExportTable.html#method-alert_warning}{\code{exceed.pipelines::ExportTable$alert_warning()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="exceed.pipelines" data-topic="ExportTable" data-id="check_study_id_range">}\href{../../exceed.pipelines/html/ExportTable.html#method-check_study_id_range}{\code{exceed.pipelines::ExportTable$check_study_id_range()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="exceed.pipelines" data-topic="ExportTable" data-id="get_field_type_overrides">}\href{../../exceed.pipelines/html/ExportTable.html#method-get_field_type_overrides}{\code{exceed.pipelines::ExportTable$get_field_type_overrides()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="exceed.pipelines" data-topic="ExportTable" data-id="get_fields">}\href{../../exceed.pipelines/html/ExportTable.html#method-get_fields}{\code{exceed.pipelines::ExportTable$get_fields()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="exceed.pipelines" data-topic="ExportTable" data-id="verify_table">}\href{../../exceed.pipelines/html/ExportTable.html#method-verify_table}{\code{exceed.pipelines::ExportTable$verify_table()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="exceed.pipelines" data-topic="ExportTable" data-id="write_dataset">}\href{../../exceed.pipelines/html/ExportTable.html#method-write_dataset}{\code{exceed.pipelines::ExportTable$write_dataset()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="exceed.pipelines" data-topic="ExportTable" data-id="write_metadata">}\href{../../exceed.pipelines/html/ExportTable.html#method-write_metadata}{\code{exceed.pipelines::ExportTable$write_metadata()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="exceed.pipelines" data-topic="ExportTable" data-id="write_table">}\href{../../exceed.pipelines/html/ExportTable.html#method-write_table}{\code{exceed.pipelines::ExportTable$write_table()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-transform"></a>}}
\if{latex}{\out{\hypertarget{method-transform}{}}}
\subsection{Method \code{transform()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExportPhenotypes$transform(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-prepare_metadata"></a>}}
\if{latex}{\out{\hypertarget{method-prepare_metadata}{}}}
\subsection{Method \code{prepare_metadata()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExportPhenotypes$prepare_metadata()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-prepare_dataset"></a>}}
\if{latex}{\out{\hypertarget{method-prepare_dataset}{}}}
\subsection{Method \code{prepare_dataset()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExportPhenotypes$prepare_dataset(metadata)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExportPhenotypes$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
