% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExportTable.R
\docType{class}
\name{ExportTable}
\alias{ExportTable}
\title{ExportTable}
\format{
An R6 class object.
}
\description{
ExportTable class
}
\details{
Generic export of a single table.
}
\section{Super classes}{
\code{\link[exceedapi:SnapshotBackend]{exceedapi::SnapshotBackend}} -> \code{\link[exceedapi:API]{exceedapi::API}} -> \code{\link[exceedapi:Step]{exceedapi::Step}} -> \code{ExportTable}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-get_fields}{\code{ExportTable$get_fields()}}
\item \href{#method-alert_warning}{\code{ExportTable$alert_warning()}}
\item \href{#method-alert_danger}{\code{ExportTable$alert_danger()}}
\item \href{#method-check_study_id_range}{\code{ExportTable$check_study_id_range()}}
\item \href{#method-add_shared_metadata}{\code{ExportTable$add_shared_metadata()}}
\item \href{#method-add_prefix}{\code{ExportTable$add_prefix()}}
\item \href{#method-get_field_type_overrides}{\code{ExportTable$get_field_type_overrides()}}
\item \href{#method-verify_table}{\code{ExportTable$verify_table()}}
\item \href{#method-write_metadata}{\code{ExportTable$write_metadata()}}
\item \href{#method-write_dataset}{\code{ExportTable$write_dataset()}}
\item \href{#method-write_table}{\code{ExportTable$write_table()}}
\item \href{#method-clone}{\code{ExportTable$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="exceedapi" data-topic="Step" data-id="add_step">}\href{../../exceedapi/html/Step.html#method-add_step}{\code{exceedapi::Step$add_step()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="exceedapi" data-topic="Step" data-id="add_step_object">}\href{../../exceedapi/html/Step.html#method-add_step_object}{\code{exceedapi::Step$add_step_object()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="exceedapi" data-topic="Step" data-id="initialize">}\href{../../exceedapi/html/Step.html#method-initialize}{\code{exceedapi::Step$initialize()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="exceedapi" data-topic="Step" data-id="inverse_transform">}\href{../../exceedapi/html/Step.html#method-inverse_transform}{\code{exceedapi::Step$inverse_transform()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="exceedapi" data-topic="Step" data-id="progress_bar">}\href{../../exceedapi/html/Step.html#method-progress_bar}{\code{exceedapi::Step$progress_bar()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="exceedapi" data-topic="Step" data-id="run">}\href{../../exceedapi/html/Step.html#method-run}{\code{exceedapi::Step$run()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="exceedapi" data-topic="Step" data-id="transform">}\href{../../exceedapi/html/Step.html#method-transform}{\code{exceedapi::Step$transform()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_fields"></a>}}
\if{latex}{\out{\hypertarget{method-get_fields}{}}}
\subsection{Method \code{get_fields()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExportTable$get_fields(exclude = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-alert_warning"></a>}}
\if{latex}{\out{\hypertarget{method-alert_warning}{}}}
\subsection{Method \code{alert_warning()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExportTable$alert_warning(message, success, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-alert_danger"></a>}}
\if{latex}{\out{\hypertarget{method-alert_danger}{}}}
\subsection{Method \code{alert_danger()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExportTable$alert_danger(message, success, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-check_study_id_range"></a>}}
\if{latex}{\out{\hypertarget{method-check_study_id_range}{}}}
\subsection{Method \code{check_study_id_range()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExportTable$check_study_id_range(study_id)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_shared_metadata"></a>}}
\if{latex}{\out{\hypertarget{method-add_shared_metadata}{}}}
\subsection{Method \code{add_shared_metadata()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExportTable$add_shared_metadata(metadata)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_prefix"></a>}}
\if{latex}{\out{\hypertarget{method-add_prefix}{}}}
\subsection{Method \code{add_prefix()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExportTable$add_prefix(dataset)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_field_type_overrides"></a>}}
\if{latex}{\out{\hypertarget{method-get_field_type_overrides}{}}}
\subsection{Method \code{get_field_type_overrides()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExportTable$get_field_type_overrides(name, type)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-verify_table"></a>}}
\if{latex}{\out{\hypertarget{method-verify_table}{}}}
\subsection{Method \code{verify_table()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExportTable$verify_table(metadata, dataset)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-write_metadata"></a>}}
\if{latex}{\out{\hypertarget{method-write_metadata}{}}}
\subsection{Method \code{write_metadata()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExportTable$write_metadata(metadata, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-write_dataset"></a>}}
\if{latex}{\out{\hypertarget{method-write_dataset}{}}}
\subsection{Method \code{write_dataset()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExportTable$write_dataset(dataset, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-write_table"></a>}}
\if{latex}{\out{\hypertarget{method-write_table}{}}}
\subsection{Method \code{write_table()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExportTable$write_table(metadata, dataset)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExportTable$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
