% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenericExport.R
\docType{class}
\name{GenericExport}
\alias{GenericExport}
\title{GenericExport}
\format{
An R6 class object.
}
\description{
GenericExport class
}
\details{
Generic export class that includes password-protected compression (optional)
}
\section{Super classes}{
\code{\link[exceedapi:SnapshotBackend]{exceedapi::SnapshotBackend}} -> \code{\link[exceedapi:API]{exceedapi::API}} -> \code{\link[exceedapi:Step]{exceedapi::Step}} -> \code{GenericExport}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{GenericExport$new()}}
\item \href{#method-get_config}{\code{GenericExport$get_config()}}
\item \href{#method-show_config}{\code{GenericExport$show_config()}}
\item \href{#method-show_summary}{\code{GenericExport$show_summary()}}
\item \href{#method-get_version}{\code{GenericExport$get_version()}}
\item \href{#method-get_source_path}{\code{GenericExport$get_source_path()}}
\item \href{#method-generate_password}{\code{GenericExport$generate_password()}}
\item \href{#method-prepare_export}{\code{GenericExport$prepare_export()}}
\item \href{#method-export_tables}{\code{GenericExport$export_tables()}}
\item \href{#method-create_encrypted_archive}{\code{GenericExport$create_encrypted_archive()}}
\item \href{#method-create_unencrypted_archive}{\code{GenericExport$create_unencrypted_archive()}}
\item \href{#method-calulate_checksums}{\code{GenericExport$calulate_checksums()}}
\item \href{#method-make_filename}{\code{GenericExport$make_filename()}}
\item \href{#method-write_csv}{\code{GenericExport$write_csv()}}
\item \href{#method-create_archive}{\code{GenericExport$create_archive()}}
\item \href{#method-create_manifest}{\code{GenericExport$create_manifest()}}
\item \href{#method-create_manifest_csv}{\code{GenericExport$create_manifest_csv()}}
\item \href{#method-create_manifest_pdf}{\code{GenericExport$create_manifest_pdf()}}
\item \href{#method-create_codebook}{\code{GenericExport$create_codebook()}}
\item \href{#method-render_markdown}{\code{GenericExport$render_markdown()}}
\item \href{#method-clone}{\code{GenericExport$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="exceedapi" data-topic="Step" data-id="add_step">}\href{../../exceedapi/html/Step.html#method-add_step}{\code{exceedapi::Step$add_step()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="exceedapi" data-topic="Step" data-id="add_step_object">}\href{../../exceedapi/html/Step.html#method-add_step_object}{\code{exceedapi::Step$add_step_object()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="exceedapi" data-topic="Step" data-id="inverse_transform">}\href{../../exceedapi/html/Step.html#method-inverse_transform}{\code{exceedapi::Step$inverse_transform()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="exceedapi" data-topic="Step" data-id="progress_bar">}\href{../../exceedapi/html/Step.html#method-progress_bar}{\code{exceedapi::Step$progress_bar()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="exceedapi" data-topic="Step" data-id="run">}\href{../../exceedapi/html/Step.html#method-run}{\code{exceedapi::Step$run()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="exceedapi" data-topic="Step" data-id="transform">}\href{../../exceedapi/html/Step.html#method-transform}{\code{exceedapi::Step$transform()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GenericExport$new(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_config"></a>}}
\if{latex}{\out{\hypertarget{method-get_config}{}}}
\subsection{Method \code{get_config()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GenericExport$get_config()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-show_config"></a>}}
\if{latex}{\out{\hypertarget{method-show_config}{}}}
\subsection{Method \code{show_config()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GenericExport$show_config()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-show_summary"></a>}}
\if{latex}{\out{\hypertarget{method-show_summary}{}}}
\subsection{Method \code{show_summary()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GenericExport$show_summary()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_version"></a>}}
\if{latex}{\out{\hypertarget{method-get_version}{}}}
\subsection{Method \code{get_version()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GenericExport$get_version()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_source_path"></a>}}
\if{latex}{\out{\hypertarget{method-get_source_path}{}}}
\subsection{Method \code{get_source_path()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GenericExport$get_source_path()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-generate_password"></a>}}
\if{latex}{\out{\hypertarget{method-generate_password}{}}}
\subsection{Method \code{generate_password()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GenericExport$generate_password(length = 32)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-prepare_export"></a>}}
\if{latex}{\out{\hypertarget{method-prepare_export}{}}}
\subsection{Method \code{prepare_export()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GenericExport$prepare_export(domain)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-export_tables"></a>}}
\if{latex}{\out{\hypertarget{method-export_tables}{}}}
\subsection{Method \code{export_tables()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GenericExport$export_tables(tables, identities)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-create_encrypted_archive"></a>}}
\if{latex}{\out{\hypertarget{method-create_encrypted_archive}{}}}
\subsection{Method \code{create_encrypted_archive()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GenericExport$create_encrypted_archive(archive, files, password)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-create_unencrypted_archive"></a>}}
\if{latex}{\out{\hypertarget{method-create_unencrypted_archive}{}}}
\subsection{Method \code{create_unencrypted_archive()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GenericExport$create_unencrypted_archive(archive, files)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-calulate_checksums"></a>}}
\if{latex}{\out{\hypertarget{method-calulate_checksums}{}}}
\subsection{Method \code{calulate_checksums()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GenericExport$calulate_checksums(file)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-make_filename"></a>}}
\if{latex}{\out{\hypertarget{method-make_filename}{}}}
\subsection{Method \code{make_filename()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GenericExport$make_filename(prefix, suffix)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-write_csv"></a>}}
\if{latex}{\out{\hypertarget{method-write_csv}{}}}
\subsection{Method \code{write_csv()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GenericExport$write_csv(x, file, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-create_archive"></a>}}
\if{latex}{\out{\hypertarget{method-create_archive}{}}}
\subsection{Method \code{create_archive()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GenericExport$create_archive(tables)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-create_manifest"></a>}}
\if{latex}{\out{\hypertarget{method-create_manifest}{}}}
\subsection{Method \code{create_manifest()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GenericExport$create_manifest(tables)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-create_manifest_csv"></a>}}
\if{latex}{\out{\hypertarget{method-create_manifest_csv}{}}}
\subsection{Method \code{create_manifest_csv()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GenericExport$create_manifest_csv(tables)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-create_manifest_pdf"></a>}}
\if{latex}{\out{\hypertarget{method-create_manifest_pdf}{}}}
\subsection{Method \code{create_manifest_pdf()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GenericExport$create_manifest_pdf(tables)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-create_codebook"></a>}}
\if{latex}{\out{\hypertarget{method-create_codebook}{}}}
\subsection{Method \code{create_codebook()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GenericExport$create_codebook(tables)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-render_markdown"></a>}}
\if{latex}{\out{\hypertarget{method-render_markdown}{}}}
\subsection{Method \code{render_markdown()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GenericExport$render_markdown(input, output_file, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GenericExport$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
