% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadSurveyResponses.R
\docType{class}
\name{LoadSurveyResponses}
\alias{LoadSurveyResponses}
\title{LoadSurveyResponses}
\format{
An R6 class object.
}
\description{
LoadSurveyResponses class
}
\details{
Load survey responses from REDCap questionnaires.
}
\section{Super classes}{
\code{\link[exceedapi:SnapshotBackend]{exceedapi::SnapshotBackend}} -> \code{\link[exceedapi:API]{exceedapi::API}} -> \code{\link[exceedapi:Step]{exceedapi::Step}} -> \code{LoadSurveyResponses}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{LoadSurveyResponses$new()}}
\item \href{#method-transform}{\code{LoadSurveyResponses$transform()}}
\item \href{#method-get_responses}{\code{LoadSurveyResponses$get_responses()}}
\item \href{#method-clone}{\code{LoadSurveyResponses$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="exceedapi" data-topic="Step" data-id="add_step">}\href{../../exceedapi/html/Step.html#method-add_step}{\code{exceedapi::Step$add_step()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="exceedapi" data-topic="Step" data-id="add_step_object">}\href{../../exceedapi/html/Step.html#method-add_step_object}{\code{exceedapi::Step$add_step_object()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="exceedapi" data-topic="Step" data-id="inverse_transform">}\href{../../exceedapi/html/Step.html#method-inverse_transform}{\code{exceedapi::Step$inverse_transform()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="exceedapi" data-topic="Step" data-id="progress_bar">}\href{../../exceedapi/html/Step.html#method-progress_bar}{\code{exceedapi::Step$progress_bar()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="exceedapi" data-topic="Step" data-id="run">}\href{../../exceedapi/html/Step.html#method-run}{\code{exceedapi::Step$run()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LoadSurveyResponses$new(
  pipeline,
  parse_factors = TRUE,
  parse_survey_fields = TRUE,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-transform"></a>}}
\if{latex}{\out{\hypertarget{method-transform}{}}}
\subsection{Method \code{transform()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LoadSurveyResponses$transform(.data, .collect, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_responses"></a>}}
\if{latex}{\out{\hypertarget{method-get_responses}{}}}
\subsection{Method \code{get_responses()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LoadSurveyResponses$get_responses(project, .collect, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LoadSurveyResponses$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
